Clicking 'move' on a character will move their character on the map until another character is selected for movement
Modified characters.ejs to call a utility function to generate an <img> element in the DOM to easily access character images for display
Update MapHelper.js to grab locX and locY from the saved X/Y location stored on each character
Clicking on the 'move' button next to each character selects them by the UID and indicates that any clicks on the map should move the avatar. This is currently done locally and not yet written back to the DB and progress will be lost when window is refreshed
Make use of game loop (currently drawing every 200ms) to do update for first iteration. Later, we will need to either have all clients query for location updates on this loop, or pass an update to all other clients when a value is updated