Add basic movement functionality for combat mode
Using a 'scale' number, used several util functions to adjust and modify the actual location on the canvas vs the stored location. ie: 5,5 on the character would be translated to something else on the world display coordinatess
Known issue: if 2 or more players are on the same 'tile' they will overlap. thinking of introducing some 'variance' number that will shift their locations depending on who is on the tile, but not a big priority right now
Another known potential issue: While we can draw a grid and simulate a grid, there is no collision detection or anything like that built in. in addition drawing avatars of different size is not yet supported