Implement re-usage of image from character profile, add bound checking to util_adjustAndCenterDisplayForCombat
Image was formerly being grabbed, and being inserted as a img tag (thus duplicating). Removed that + the helper function, and instead used jquery to grab the element by the attribute, grab the child img element and use that instead for display.
Bounds checking was not really being done by the util, therefore allowing a negative input. Adding in a conditional return so that anything less than 0 would bump it back to the center of the 'tile'