Remove hardcoded locations, create function to draw mapview, fix issue with rendering when no profile pic is supplied, create a populate test mode
Hardcoded locations in MapHelper made it impossible for anybody else to test. Added a test function to generate a set of test data no matter what room this is created in.
Without uploading an image, last commit would cause a crash when it tried to render a character without a photo. Added in a condition so it would draw a default instead.
Added a function to trigger a redraw of the mapview. Adding a character will now also cause it to render right away.
Some formatting